pyproject.toml
[project]
name = "procugents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
]

[tool.uv.workspace]
members = [
    "dashboard",
]

[dependency-groups]
dev = [
    "langchain-anthropic>=1.4.1",
    "langchain-core>=1.3.2",
    "langchain-openai>=1.2.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
